requirements.txt
Flask
Flask-Mail
Flask-Login
Flask-WTF
Flask-SQLAlchemy
email-validator
pymysql
python-dotenv
stripe